backend/core/requirements.txt
#msal
gunicorn
Flask>=2.2
requests>=2,<3
ms_identity_python[flask] @ https://github.com/azure-samples/ms-identity-python/archive/refs/heads/0.9.zip
python-dotenv<0.22